requirements.txt
python-dotenv
openai
azure-storage-blob
psycopg2
sentence-transformers
requests